+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Versions follows PEP440 version scheme.
 
 Fixed
 =====
-
+* 'Event Title' changes in the admin site immediately reflect in the user dashboard chart
 * fix: [RGG-591] badges cannot be configured with the pre-condition depend on itself
 * admin permissions are required to access Badge API
 * View on site button in the admin panel on the StatusBadge edit page opens correct URL for image